game-commit-0003_WinLose logic implemented
