+++ index.html
@@ -55,15 +55,19 @@
             <div class="outfit-container">
                 <div id="outfitImage" class="outfit-image"></div>
             </div>
-            <div class="end-screen">
-                <div class="win1-screen"></div>
-                <div class="win2-screen"></div>
-                <div class="win3-screen"></div>
-                <div class="lose1-screen"></div>
-                <div class="lose2-screen"></div>
-                <div class="lose3-screen"></div>  
-            </div>
         </div>
+    </div>
+    <div class="end-screen">
+        <div class="notouch-screen"></div>
+        <div class="win-screen">
+            <div class="win-text">YOU WIN!</div>
+            <div class="win-image"></div>
+        </div>
+        <div class="lose-screen">
+            <div class="lose-text">LOSER!</div>
+            <div class="lose-image"></div>
+        </div>
+        <div id="homeButton2" class="home-button2">HOME</div>
     </div>
     <script src="js/main.js" type="module"></script>
     <script src="js/room.js"></script>
